chain: Move BlockStreamMetrics into block_stream module
